feat: Add new widgets and factory functions to TUIKit
- Introduced new widgets: TUITreeView, TUITextArea, TUIProgressBar, TUIScrollableContainer.
- Added widget and layout aliases for easier usage.
- Implemented factory functions for new widgets to streamline widget creation.
- Updated TUIStyle to remove the Default theme and refactor theme initialization.
- Fixed potential out-of-bounds access in TUIComboBox, TUIMenu, and TUIRadioBox by using size_t for index comparisons.
- Refactored TUIButton to correct the order of member variables.
- Enhanced TUIComboBox to track the last known index for better state management.
- Cleaned up constructor parameters in TUITextField and TUIWidget.

+++ src/core/TUIApp.cpp
@@ -154,7 +154,7 @@ int TUIApp::exec() {
         if (show_modal_) {
             return modal_->OnEvent(event);
         } else {
-            if (!pages_.empty() && active_page_ >= 0 && active_page_ < pages_.size()) {
+            if (!pages_.empty() && active_page_ >= 0 && static_cast<size_t>(active_page_) < pages_.size()) {
                 ftxui::Component active_page_component = pages_[active_page_];
                 if (active_page_component->OnEvent(event)) {
                     return true;
@@ -198,7 +198,7 @@ void TUIApp::add_page(const std::string& name, ftxui::Component page) {
 }
 
 void TUIApp::set_active_page(int index) {
-    if (index >= 0 && index < pages_.size()) {
+    if (index >= 0 && static_cast<size_t>(index) < pages_.size()) {
         active_page_ = index;
     }
 }
