+++ src/widgets/TUIMenu.cpp
@@ -31,14 +31,14 @@ int TUIMenu::selectedIndex() const {
 }
 
 std::string TUIMenu::selectedText() const {
-    if (selected_index_ >= 0 && selected_index_ < options_.size()) {
+    if (selected_index_ >= 0 && static_cast<size_t>(selected_index_) < options_.size()) {
         return options_[selected_index_];
     }
     return "";
 }
 
 void TUIMenu::setSelectedIndex(int index) {
-    if (index >= 0 && index < options_.size()) {
+    if (index >= 0 && static_cast<size_t>(index) < options_.size()) {
         selected_index_ = index;
         // Force update of the FTXUI component to reflect the new selection
         update_ftxui_menu_component();
